Event management

+++ Engine/Engine/Core/Timer.cpp
@@ -10,7 +10,7 @@ void Timer::update()
 {
 	prev = curr;
 	curr = clock();
-	dt = (curr - prev) / (float)CLOCKS_PER_SEC;
+	dt = (float)(curr - prev) / (float)CLOCKS_PER_SEC;
 	t += dt;
 	fps = 1 / dt;
 }
